move action out to new module
